requirements.txt
requests>=2.31.0
fastmcp>=0.5.0
python-dotenv>=1.0.0